Add basic say editor with debug rendering
This is the basic say editor with all additions necessary to make it
render, including the latest changes of @abeforgit.

This renders a standard editor with nothing in the sidebar.  It lets you
edit the content and the raw html content is emitted at the bottom of
the editor for quick debugging.